Added navbar and more CSS classes
Navbar is added with styling. Colors need to be fixed because I just copied the styling from W3Schools (I'm a lazy programmer. deal with it.) No links yet because no pages.

CSS classes exist for formatting rows of elements in an easy way. Sizes exist for one-half of the row, one-third, one-quarter, and one-sixth. One-eighth may be added if necessary.