Made the requested changes, note the navbar WILL utilize root-relative paths. As will <link> tags. It's the easiest that way.
For the record, if it is acceptable for the navbar to use root-relative paths, why, exactly, can I not use it to reference the BPS Robotics website? Or 1510?

If I'm going up one level in the 2898 site, it makes sense to use pwd-relative paths (for example, on a forum go back to the forum from the thread you're on), but if I'm going somewhere that honestly has nothing to do with the 2898 site (like linking to the 1510 site) why would root-relative paths be wrong?